requirements.txt
certifi==2019.11.28
chardet==3.0.4
idna==2.8
progressbar2==3.47.0
python-utils==2.3.0
requests==2.22.0
six==1.13.0
spotipy==2.4.4
urllib3==1.25.7
